Replace SDL_AudioStream with the linear resampler.
The quality will definitely be lower, but the problem with the SDL resampler is
that it introduces padding frames that we don't actually care about. After a
while, it starts to hoard samples and eventually the submix falls behind, making
a strange and extremely subtle audio lag that can build up over several hours.