Release the source list mutex when processing sources.
Client callbacks may have locks that depend on another thread doing
XAudio work, so if for example the audio thread is in the middle of an
OnBufferEnd call while the main thread is trying to create a source, the
main thread will wait for the source processing to end, but the audio
thread will be stuck in the client's mutex! Puyo Puyo Tetris is one
example of this.