Enforce a quantum size measured in milliseconds, overriding the OS buffer size.
The spec states that the default update interval is 10ms, with 1024_QUANTUM
being a somewhat poorly-named flag to use a 21.33ms interval instead. So, the
graph needs to be consistent in size no matter what the OS wants, we'll need to
depend on SDL doing the right thing with the internal SDL_AudioStream.