Make sure the OnBufferEnd events from FlushSourceBuffers are called from the mixer thread.
As tests demonstrate, FlushSourceBuffers is asynchronous (at least with
XAudio 2.7), and doesn't immediately call OnBufferEnd from whatever thread
it was called from (even if it was from a callback). Instead, they are later
flushed from the mixer thread with the OnBufferEnd events sent accordingly.

Note that GetState's BuffersQueued also counts the buffers that haven't
been removed yet, even if they are marked for deletion.

The game Legend of Heroes: Trails of Cold Steel II (steam appid 748490)
relies on both of these behaviors, else it hangs on exit.

Signed-off-by: Gabriel Ivăncescu <gabrielopcode@gmail.com>