Add support for client-provided platform callbacks.
Even with the engine procedure callback, Wine has to swap it to another
thread in order for the application-provided callbacks to be executed in
the context of a Wine thread, and not the SDL audio created thread.

This adds unnecessary complexity, where we can simply provide some
platform callbacks to create threads and synchronization primitives
directly instead.

This will also allow Wine get rid of the FAudio SDL2 dependency, which
is desirable, considering we'll ultimately need FAudio built as PE.