For [X]WMA, send decoded packet size to the decoder
This patch sucks a little bit. There doesn't seem to be any standard way
to communicate the DPDS data to decoders via gstreamer.

GstAudioClippingMeta comes close, but it communicates how many samples
to trim from the resulting buffer, not how many samples we expect to be
decoded. Another option may be some form of seeking (this is how ffmpeg
consumes it when reading directly from a file), but I couldn't find a
purely informational way to communicate a seek table that wouldn't
result in some functional change.

So the best I could come up with is a custom GstMeta, which only one
decoder, Proton's media-converter, will actually consume. Not great, but
it shouldn't interfere with any other decoder.