build: fix dependency on SDL2
This is a followup on 81d5a24 (build: fix building with MSVC,
2024-03-07) and 68a555e (build: fix __attribute__((aligned(x))) on
gcc/clang, 2024-03-09) for the purposes of creating a vcpkg FAudio port.

Add a `Requires.private: sdl2` to the pkgconf file for the SDL2
dependency when not using PLATFORM_WIN32.

Add the SDL2 dependency to the generated cmake config file when not
using PLATFORM_WIN32 as well.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>