[windows] List all audio devices for Windows
Previously, FAudio with `PLATFORM_WIN32` would only list the default
audio device. This reworks the Faudio `PLATFORM_WIN32` code to enable
access to every audio device on the system. In order to preserve the
existing behavior (and Xaudio2 compabitility), the audio devices are
re-ordered so the first device is always the default one.

In addition, this properly populates the `DisplayName` field in
`FAudioDeviceDetails`. Previously, it was set to the device GUID.

Test: Local build with visualboyadvance-m