build: separate shared and static import targets
Export the FAudio::FAudio-shared target when building a shared library
and the FAudio::FAudio-static target when building a static library to
separate target files so that both can be installed at the same time.

Alias FAudio::FAudio to the shared library if it is installed and the
static library otherwise in the Config.

Signed-off-by: Rafael Kitover <rkitover@gmail.com>