FACT: Queue wavebanks instead of complete PREPARED notifications.
Notifications are supposed to be generated when they are sent. That is,
registering a notification between preparing a wavebank and calling DoWork()
will still result in the notification being sent.

Therefore we cannot set the notification's time or context when preparing the
wavebank, and must delay until DoWork(). This also reduces the amount of space
we are using.