FACT: Create an FACTSoundInstance immediately if possible.
And set its state to a new state SOUND_STATE_STOPPED.

For non-interactive cues, GetProperties() is supposed to be able to retrieve cue
properties for the sound and track variations which will play, even before
Play() has been called.

Achieve this by initializing the sound instance earlier, but don't actually play
it yet. Instead we use a new state to signify that the sound is not stopped.
This state exists in complement to the "fading in" and "fading out" states, and
is essentially only used to affect when the FAudio sound is played.